requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
deepdiff==6.7.1
distro==1.9.0
emoji==2.10.1
faiss-cpu==1.8.0
filetype==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain-community==0.0.28
langchain-core==0.1.31
langdetect==1.0.9
langsmith==0.1.25
lxml==5.1.0
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.14.0
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pydantic==2.6.4
pydantic_core==2.16.3
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
